Refactored TransactionTableViewCell
Didn’t like the fact that my last change (adding formatter) meant that
the cell’s dependencies had to be set in the correct order or it
wouldn’t work.  Therefore created a ‘set content’ function that is
effectively a kind of initialiser.  Also noticed that cell doesn’t need
to keep a reference to Transaction so did away with that dependency
altogether.