Continued work on Reports views
Made the date pickers ‘collapse’ as is standard in iOS, also disable
generate button when date range is invalid (i.e. end date is before
start date).
Changed report list to have optional reportData.  This allows for
reportData to be nil before any report is generated, in which case the
screen is ‘blank’, whereas when the report is empty we show that with
an ‘empty’ cell.