Added ReportTransactionDetailsTableViewController
This shows the individual transactions for a category in a report.  In
order to reuse the TransactionsDataSource class I had to remove its’
account property and instead tell it what its’ data is.  Apart from
this is was simple to create the new class, as can be seen by how short
its’ code is.